Pulled out the actions into their own file for ease of use
